clean up invocation extensions, move Command.Invoke* to System.CommandLine namespace
