+++ src/System.CommandLine.Tests/ParseDirectiveTests.cs
@@ -43,7 +43,7 @@ namespace System.CommandLine.Tests
 
             var console = new TestConsole();
 
-            await parser.InvokeAsync(result, console);
+            await result.InvokeAsync(console);
 
             console.Out
                    .ToString()
